Fix tests related to worklog #5370
http://bazaar.launchpad.net/~mysql/mysql-server/mysql-5.1/revision/3484
was done for http://forge.mysql.com/worklog/task.php?id=5370. It
changed the code so that when CREATE TABLE IF NOT EXISTS ... SELECT
... is run for a table which does exist 2 statements are put into the
binlog. The slave treats the 2 statements as separate groups but the
stock MySQL code was putting both statements into the transaction
cache, which confused the group_id code. The code was fixed in the
group_id commit so that for SBR the master puts the events in
separate groups. This commit fixes the related tests.

Change-Id: Ic2569448c6f5160724ce52249003b6c9a609228c