Extract test for bug 56226 to its own test
The test for bug 56226 used to be done as part of the rpl_alter test.
Since the test forces binlog format to row this led to duplication
of effort because it ran for all rpl combinations. Extract it out to
its own test file so that it only runs for the 'row' combination.

Change-Id: Ic1b2c8c3c299eb6ad63b17418542710885395a0c