Extract test for bug 55625 to its own test
The test for bug 55625 used to be done as part of the rpl_drop test.
Since the test forces binlog format to row this led to duplication of
effort because it ran for all rpl combinations. Extract it out to its
own test file so that it only runs fors for the 'row' combination.

Change-Id: Ia9e6c962132cdffec9a57269fa8d11b89a017d07