rpl.rpl_extra_col_master_* only works for RBR
The rpl.rpl_extra_col_master_* tests as duplicating work because
they are explicitly setting the binlog format mutliple times even
though the rpl combinations file already does this. In addition,
the test is currently not testing what was intended. See
http://bugs.mysql.com/bug.php?id=60542. Per that bug, the problem
is fixed in 5.5 so copy the test's files from the 5.5 tree.

Change-Id: I1f6d923b7c5a1bd839a4bb0030312c56ced361f4