Fix race between Binlog Dump threads and RESET MASTER
Change RESET MASTER so that it kills all Binlog Dump threads before
deleting logs. This fixes a race between the Binlog Dump threads
calling close on a file descriptor which was for a file which was
deleted out from under the thread and any other thread which may have
called open or create and gotten back that same file descriptor.

Change-Id: Ice194223b8d55a3fdc93cc4824fc95c27e595c29