Add proxy authentication
Using the --proxy-user=[user] command line option, a MySQL user can
be given the ability to execute COM_CHANGE_USER without providing
the credentials of the user to be changed to. This is a useful
solution to credential management, as a proxy can authenticate as
a user externally and then perform COM_CHANGE_USER to create a
connection as the desired user.

Change-Id: I2d7ed8105eb0b9fb5f86bd6d8bc75c61898f739f