Fix for rpl.rpl_crash_on_binlog_io_error for --rpl-hierarchical
When --rpl-hierarchical is used the crash recovery procedure involves
reading the binary log. The rpl.rpl_crash_on_binlog_io_error test is
intentionally corrupting the binlog to crash the servers. When the
server comes back it hits an additional error. This patch is adding
this error to the list of suppressed messages.

Change-Id: I347f04dda937a512868dcbb8fd392625b3ae8223