Initial patch porting table statistics from 5.0
This is the initial patch to port table stats from 5.0 to 5.1.
It includes the storage and display of the stats but does not
include the code to actually update them.

The implementation is a bit different from 5.0 because I hooked
into the code new to 5.1 which combines SHOW commands with the
INFORMATION_SCHEMA so that the stats can be obtained either way.

Entries do not show up in the table statistics until either DML
is run to update rows or a SELECT done to read them. Additionally,
for temporary tables, the stats are not updated until the table is
dropped, either explicitly or implicitly when the connection is
closed. Note that dropping a table does not clear the stats.

There have been some changes compared to 5.0. In 5.0, the output
looked like:

mysql> SHOW TABLE_STATISTICS\G
*************************** 1. row ***************************
                  Table: scratch.BillingTypes.root
              Rows_read: 0
           Rows_changed: 2
Rows_changed_x_#indexes: 2
                 Engine: InnoDB
1 row in set (0.00 sec)

In 5.1 that same output is now:

mysql> SHOW TABLE_STATISTICS\G
*************************** 1. row ***************************
                  Table: BillingTypes.root
                     Db: scratch
              Rows_read: 0
           Rows_changed: 2
Rows_changed_x_#indexes: 2
1 row in set (0.00 sec)

and in 5.1 can be accessed via INFORMATION_SCHEMA:

mysql> SELECT * FROM INFORMATION_SCHEMA.TABLE_STATISTICS\G
*************************** 1. row ***************************
             TABLE_NAME: BillingTypes.root
           TABLE_SCHEMA: scratch
              ROWS_READ: 0
           ROWS_CHANGED: 2
ROWS_CHANGED_X_#INDEXES: 2
1 row in set (0.00 sec)

Change-Id: Ie3a9e0019d92e73bb2429f435fafea08d225818c