Initial patch porting user statistics from 5.0
This is the initial patch to port user stats from 5.0 to 5.1.
It includes the storage and display of the stats but does not
include the code to actually update them.

The implementation is a bit different from 5.0 because I hooked
into the code new to 5.1 which combines SHOW commands with the
INFORMATION_SCHEMA so that the stats can be obtained either way.

Change-Id: Ie55ff5d482f18ed93ca92cd62d6777c32dd0926d