Another fix for rpl.rpl_crash_on_binlog_io_error
This test did not appear to be functioning as intended as the
backticks seemed to be sending the select to the shell rather
than mysql so $new_pos wasn't getting set correctly. Fix test
so that the master actually hits an error on the test. Also,
fix the suppressed error regex so that the parens get matched
correctly.

Change-Id: I96085f23340502eec4fb3f4f4e1c573342c7ce1b