Part I of porting mapped user support to 5.1
Initial patch adding mapped user support to 5.1. Mapped users are
lightweight users; these are accounts that have a password but no
privileges. They map to one regular account which provides the
privileges and should be treated as a role. A new system table,
mapped_user, is added for this.

This patch introduces the following new commands:

  CREATE MAPPED USER user [IDENTIFIED BY [PASSWORD] 'password'] ROLE role;
  DROP MAPPED USER user;

Change-Id: I6fb03f935872e647a2ffb27f54e11475c23a5407