Add ROLE column to PROCESSLIST and USER_STATISTICS
Add a ROLE column to the PROCESSLIST and USER_STATISTICS tables
of INFORMATION_SCHEMA. For mapped users this new column obviously
contains the role account from which it gets its privileges. If
the user is not a mapped user then the column contains the same
value as the USER column.

Note that this new column also shows up in SHOW USER_STATISTICS
but does not show up in SHOW PROCESSLIST. The latter has a
separate implementation from the INFORMATION_SCHEMA table
which I didn't update in order to be backward compatible.

Change-Id: I810e275b2d9d062d72da7e28ba8b0c5fe3763015