Various index/optimization repairs/workarounds for GoogleStats
opt_range.cc: fix an infinite loop on condition pushdown when the last
range is filtered by the condition. This is actually something of a
deoptimization, but is safer than attempting to repair the underlying
bits of the optimizer.

opt_sum.cc: do not use prefix_last_scan (used by MAX()) if the
underlying storage engine does not support it.

Change-Id: Id995e20c72f5b5e76bd766820a054ba9ea9bdfe0