Infrastructure to support the SQL log
Add a new commandline / my.cnf variable --sql-log. It takes
an optional string parameter which is the base name of the log.
Actual files written to disk are constructed from the base
name with a sequential numeric value appended. This patch
also includes the support for SQL log handling of FLUSH LOGS.

Change-Id: I3010f1b994308263b5e30a0b220233bf54537db4