Add the IEEE754_TO_STRING function
The conversion from double -> decimal -> double is lossless when the
conversion from double -> decimal generates 17 digits of precision.
The function IEEE754_TO_STRING makes that possible.

This patch also adds an option, --lossless-fp, to mysqldump which
converts double and float to decimal with extra precision so the
reinserted values will be equal to the original values.

Change-Id: Iee40a397872daf163e36a96ebbb5dbbe35a2b975