Another rpl.rpl_hier_crash_recovery fix
In my last fix for an intermittent failure I ended up introducing
a different intermittent failure. There is a race in mysql-test-run.pl
which hits when a test shuts down multiple instances of mysqld.
Sometimes it will end up looping waiting for the wrong instance
to restart, causing the test to fail.

Change-Id: I22dea13dfe0638bc038d803f578bbe4329398cca