Unit tests for SQL logging
This change adds unit tests for SQL logging. Some of the tests
are restricted to only running with statement-based replication
as the logging is not currently correct when using row-based
replication. I will fix that in a follow-up change.

Also, there has been a change between 5.0 and 5.1. In 5.0 if
an UPDATE statement didn't actually result in any changed data
on a row it was still written to the SQL log. That is no longer
the case in 5.1 due to the fix for
http://bugs.mysql.com/bug.php?id=29157. Thus, the result file
for the rpl.rpl_sqllog_inno test is different from 5.0.

Change-Id: Ibc41596519f9a912b6c77377ae0b5da422d8548c