Increase the /var size in httpd-stress
This change improves the ./google/httpd-stress by increasing the size of
the /var page. This is done by creating many tables and then querying
them using many users. The duration of test was also capped to a
configurable number of seconds. The tests can be run with a bigger
number of tables and users by adding 'long' as a second parameter to the
./google/httpd-stress (the total durations bumps up to about 22
minutes).

Change-Id: I62b8f6e1792b2b5e22b6b1823106657450128f2e