Return error to clients on semi-sync repl failures
This patch adds error reporting of semi-sync failures, when the
connection was killed before the master could receive an ACK from one of
the slaves.
When a client receives this error, the transaction has been committed to
the master, but not necessarily to any of the slaves.

If this happens, it could mean that the master was already replaced, and
that this transaction is an alternate future. Manual cleanup is required
in this scenario. The aim of this patch is to propagate the error up to
the clients, in the hopes that it is noticed.

Added some grammatical corrections in this patch as well.

Change-Id: Iff428cedc9e2861a6d35e190bfafb1e49d2eb988