Add support to select only one test combination
This change adds a new option (--one-combination=s) to mysql-test-run.pl
which makes it possible to run only one of the combinations defined in a
combinations file. The --suite should be used to pick a certain
combinations file.

What I was trying to do is running each combination on a different
machine. One thing I tried is taking the settings from the combinations
file and pass them via the --combinations flag but that doesn't work
because --combinations doesn't support multiple flag changes. I also
tried doing the same using --mysqld but that doesn't interact very well
with the rather convoluted filtering of flags.

Change-Id: I91c122865530108f42150be902790f9e3aae7cf3