Fix semi-sync deadlock
In the patch to harden semi-sync masters against bad semi-sync
packets a deadlock was introduced because different threads
were both acquiring mysql_bin_log.LOCK_log and semi_sync_replicator.lock
but were doing so in opposite orders. Change the harden fix to
acquire the locks in the other order.

Change-Id: I67812c56c391b858881ba3455b50103a3644ff6f
Reviewed-by: Razvan Musaloiu-E. <razvanm@google.com>