Stop memory blow out in get_quick_keys()
For certain types of input (e.g; multi-thousand item IN lists of
unsorted content), get_quick_keys() recurses on itself thousands and
thousands of times. We add a simple counter to count these recursions,
if we get in too far, we bottom out the recursion which tells the
optimizer to do something else.

Change-Id: Iedbf6393adf65f66eb177015b60bb900c7b0a2de
Reviewed-by: Pavel Ivanov <pivanof@google.com>