Clean up test compilation targets
Older way of doing things made it difficult to see the source file of
the test while debugging in gdb/cgdb. Also copying/removing sources
inside of source directory at compile time is not really safe.

This commit also fixes a few compilation problems for the tests, that
were happening due to bitrot and lack of love. The tests in strings/ now
compile, however they do not link and that should be fixed separately.

Change-Id: Ia289fd779a6fb826794c31f6fd4e519824bd3613
Reviewed-by: Andrew Ford <andrewford@google.com>