Add special system users support.
System users are stored in mysql.system_user table which has identical
structure as mysql.user table. These users behave the same as regular users
during authentication except that they are preferred over any user in mysql.user
table (it matters not only when mysql.user has record with the same user name
and host name but when e.g. mysql.user has record for ''@'192.168.0.1',
mysql.system_user has record for 'user'@'192.168.%' and authentication is
attempted for 'user'@'192.168.0.1'). System users also cannot be created,
changed or dropped using regular CREATE/DROP USER commands or GRANT command with
user-level permissions. The only way to manage these users is through direct
modification of mysql.system_user table with the following "flush privileges".
And that can be done only by users with SUPER privilege.

Also add an is_system_user boolean to the Security_context structure, which is
set to true if the current user was loaded from the mysql.system_user table.
In order to ensure that the setting of this flag is testable, and in case it is
useful to the user, it is also exposed as a function, IS_SYSTEM_USER().

The IS_SYSTEM_USER() function returns 0 if the user ACL was loaded from
the normal mysql.user table, and returns 1 if the user was loaded from
the mysql.system_user table.

System users functionality is turned on with --system-user-table command line
flag.

Change-Id: Ibed77b90f2716b50adc1fac59f5a8dc844480a88