Remove mysql tables from consideration by innodb-zip test
The main.mysqlcheck test optimizes the tables in the mysql database,
which makes this test's output dependent on whether that test has run
previously in the same worker.  By only examining a test-specific
database in its information_schema checks, the new test should be more
robust to being run in varying starting conditions, including if some previous
test ever fails to clean up some of its tables.  This is in line with
test design recommendations at
http://dev.mysql.com/doc/mysqltest/2.0/en/writing-tests-cleaning-up.html

Change-Id: I6dc53b40a4aeeb9e851783cf47ff79c59060c21f
Reviewed-by: Pavel Ivanov <pivanof@google.com>