Fix crash bug in deprecated_engines change
A full test run revealed that the new ALTER TABLE code segfaults if
an ALTER TABLE query is executed with an ENGINE = clause referring
to an unrecognized storage engine.  This only turned up because ndb
was not compiled into the server for the test run, so extra queries
are added to the deprecated_engines test using ENGINE = frobnitz,
and using information_schema to verify that this engine name
remains unrecognized just in case.

Change-Id: I68e65da7bf195db9d7c253feefe422a9cd601b16
Reviewed-by: Pavel Ivanov <pivanof@google.com>