Improve accuracy of LIMIT-related replication safety checks
Before this change, any DELETE, UPDATE, or INSERT...SELECT query with a
top-level LIMIT clause was considered unsafe for statement-based
replication, since the order of rows and therefore which subset would be
selected is undefined in the general case.  However, in some cases it is
possible to establish based on the query structure and schema that the
row ordering is deterministic, making it safe to replicate statements
with LIMIT.  Also, the previous check did not consider LIMITs in
subqueries or UNION queries, both of which are potentially problematic.

In place of the previous simple check for top-level LIMIT clauses, this
change adds a new function
is_read_clause_safe_for_stmt_replication(THD*) that examines the read
path of a DELETE/UPDATE/INSERT...SELECT query to see if there are any
constructs that are unsafe for statement-based replication.  It's still
imperfect, but should be more accurate than the old behavior.

If a LIMIT clause is used anywhere, then the query and any of its
subqueries must also ORDER BY all of the columns of at least one primary
or unique+not-null key in each table involved.  UNION queries with outer
LIMIT clauses (attached union-wide or in outer queries when a UNION is a
subquery) are considered unsafe, because it is difficult to determine
whether a particular ordering is deterministic across the whole union.
This is a bit conservative.

This change also marks queries that write data derived from
information_schema tables as unsafe, LIMIT or no, since those tables may
contain replica-specific runtime data.  It should be possible in the
future to make exceptions for standard information_schema tables whose
contents are entirely determined by the physical contents of the
database and whose structure is unlikely to change.

Change-Id: I999dc4e3a4fc2d10356e225c0a3f307045a727ff
Reviewed-by: Pavel Ivanov <pivanof@google.com>