PORT: BUG#13545447: RPL_ROTATE_LOGS FAILS DUE TO CONCURRENCY ISSUES IN REP. CODE
In method mysql_binlog_send, right after detecting a EOF in the
read event loop, and before deciding if we should change to a new
binlog file there is a execution window where new events can be
written to the binlog and a rotation can happen. When reaching
the test, the function will then change to a new binlog file
ignoring all the events written in this window. This will result
in events not being replicated.

Only when the binlog is detected as deactivated in the event loop
of the dump thread, can we really know that no more events
remain. For this reason, this test is now made under the log lock
in the beginning of the event loop when reading the events.

***
This also ports a second commit for this bug that fixes a build error after just
porting the first. It's commit message was:
    Post-push fix, broken build:
    sql/rpl_master.cc:1049:70: error: converting ‘false’ to pointer type ‘bool*’ [-Werror=conversion-null]

Further, I have gone back to the original test from http://bugs.mysql.com/bug.php?id=67929
which uses sleep instead of debug_sync. The modified test with debug_sync is not available
(or at least I couldn't find it), and it's too painful getting it to work.

Change-Id: I91a5a4044056b8d3fc8c007c89b224da3683aadf
Reviewed-by: Jeremy Cole <jeremycole@google.com>