Fix for alarm thread pthread_kill issue.
This issue:
  - alarm_queue.elements > 0
  - SIGTERM (or SIQQUIT) comes in
  - kill_server_thread spawned
  - kill_server_thread: calls clean_up()
  - kill_server_thread: increments cleanup_done
  - signal_handler_thread: gets another signal
  - signal_handler_thread: checks cleanup_done and kills itself
  - kill_server_thread: calls end_thr_alarm, which calls
      pthread_kill on signal_handler_thread, which is already dead

cleanup_done does not actually indicate that the cleanup is done, it is
being used to only call clean_up once (and is even failing at that technically).
I fixed that, and made cleanup_done actually indicate that cleanup is done, so
it is safe for the signal_handler_thread to pthread_exit when that is set.

Change-Id: I58436f9a8f4cbdc64078d8d4e26c85a8b9d982bd
Reviewed-by: Ivan Tarasov <vanya@google.com>