Add support for hide_sensitive_information variable.
This SUPER-only, GLOBAL-only variable hides potentially sensitive
information such as server and client hostnames and IP addresses so
that they cannot be seen by non-SUPER clients while it is enabled.
Instead, they will see the string "(hidden)".

Note that this patch does not attempt to hide the hostname or other
sensitive information that may be present in filenames, replication
status commands, or in log files themselves, as additional privileges
should protect those from the user.

Change-Id: I4bedd5f471a8e7678fb94359b18f7ad18e5eb798
Reviewed-by: Pavel Ivanov <pivanof@google.com>