Add @@disk_quota_exceeded sysvar to mysqld
This adds a new global disk_quota_exceeded setting, and causes the
server to refuse writes when it is enabled in the same manner as
for the existing read_only setting.  read_only continues to work
as before.

Tests are copied from read_only and read_only_innodb, with only the
sysvar name and expected error messages modified.

Change-Id: I55bb3c6ad7148a3db4915d1ad197e9d61858d08d
Reviewed-by: Andrew Ford <andrewford@google.com>