Add more logging for unspecified innodb errors.
Make sure there are log messages explaining the error
when DB_ERROR is generated.

This also changes the converted error to HA_ERR_INTERNAL.

I tried to find cases where an error would not be converted
from InnoDB to MySQL, but the tests do not appear to do that at all.
It's possible that more testing is needed for that, but that case should
probably not ever happen anyway (meaning all InnoDB errors should have
a corresponding MySQL error to convert to).

Change-Id: I416b56750ce0edf20d82a039d36e49c1e7340dc2
Reviewed-by: Jeremy Cole <jeremycole@google.com>