Add restricted_schemas feature.
The restricted_schemas variable allows certain schemas (databases) to be
restricted to access by only users with SUPER or users from the
system_user table. This is intended to protect those schemas from access
by users who may otherwise have broad access or access to all schemas.

The restricted_schemas variable can be set to a comma-delimited list of
schemas to protect. The variable will be displayed in SHOW VARIABLES and
via its @@-variable in a parsed and reconstructed form based on the
internal list.

Change-Id: Iddf4e072782a79bd7dc45b74f3744a6ebed92acc
Reviewed-by: Pavel Ivanov <pivanof@google.com>