Add require_super_for_mysql_schema_ddl feature.
This feature allows non-SUPER users' DDL commands against the 'mysql'
database to be blocked while still allowing DML commands to be used.
The primary purpose is to allow "regular" users to administrate their
own permissions while not allowing them to change the structure of,
to damage, or to corrupt the permissions tables themselves.

Change-Id: I25da2049614b450dfd22f62c37d0be270f9b7d70
Reviewed-by: Pavel Ivanov <pivanof@google.com>