Add require_super_for_mysql_schema_ddl feature.
This feature allows non-SUPER users' DDL commands against the 'mysql'
database to be blocked while still allowing DML commands to be used.
The primary purpose is to allow "regular" users to administrate their
own permissions while not allowing them to change the structure of,
to damage, or to corrupt the permissions tables themselves.

Change-Id: Iaa462cea3eff93262f454f24d015765140ae8825