Add the restricted_variables feature.
The restricted_variables variable allows certain variables' values to be
hidden from non-SUPER users. When a variable is restricted, non-SUPER users
will see the string "(hidden)" instead of the variable's value in SHOW
VARIABLES and SELECT @@variable queries.

The restricted_variables variable can be set to a comma-delimited list of
variables to protect. The variable will be displayed in SHOW VARIABLES and
via its @@-variable in a parsed and reconstructed form based on the
internal list.

Change-Id: Ice40fe20dc6bf067dffb1cb5b2c55b3e91eea652