Fix flakiness of some tests.
Several issues fixed:
- Fixed comparison in my_b_read and my_b_write which could overflow in 32-bit
  compilation and give incorrect results after that.
- Added one more legal error code into rpl.rpl_killed_ddl test. Looking at other
  expected error codes it seems that "Unknown event" can also be expected.
- Fixed different ASSERTs that don't play nicely with files modified outside of
  MySQL.
- Fix rpl_gtid_basic.test flakiness by ensuring that DROP TABLE has
  propgated to all slaves before continuing on next part of test,
  which does RESET MASTER.
- rpl.rpl_gtid_mdev4820:
  Apparently master gives different error depending on whether slave connects
  before the next INSERT statement or after it. Make sure that slave is
  completely started before executing INSERT on master.
- rpl.rpl_binlog_index:
  It looks like sometimes some files are not yet deleted from temp directory
  before an attempt to delete directory is made. Just use rm to recursively delete
  all the files.

Also unflake binlog_mysqlbinlog2

binlog_mysqlbinlog2 is flaky as group commit code for mariadb
is not 100% deterministic wrt to thread scheduling. when "flush logs"
is executed the binary logs is rotated, and when the file is opened
a "binlog checkpoint" is written. This checkpoint depends on
if the storage engine (innodb) has fsync:ed it's log, something that
is not needed with the binlog group commit optimizations made.
the result is that there can sometimes be 2 binlog checkpoints
in the newly opened binlog (one created when file was opened, and
one created later when the innodb logs where fsync:ed) and sometimes
only one if innodb logs happened to get fsync:ed earlier.
this means the hard-coded binlog offsets used in the test can
happen to point to different binlog events, hence give different
output.

this patch avoids this problem by using different hard-coded offsets
that are not sensitive to the problem described above.

the original hard-coded offsets pointed into middle of an event,
which had a size less than the size of a binlog checkpoint event.
i.e the last event would be different depending in if the checkpoint
was in or not. the new hard-coded offsets points to the first
event after the first checkpoint (that will always be there).
this position is quite uninteresting, but this test is only to check
that it's possible to set a stop-position for mysqlbinlog.

Also fix flakiness of rpl_checksum by not showing binlog positions
which unfortunately depend on the background thread's binlog_checkpoints.

Change-Id: I7a126d4c273aaa0d4650b3845fb02a3827ed98ac