Fix race in hide_sensitive_information test.
SHOW VARIABLES queries can generate temporary tables which are removed after
client already recieved all results and mtr already continued on to execution
of the next statements. If at this point we execute SHOW PROCESSLIST on the
other connection we can see the first connection in the state "removing tmp
table". And that won't match our expectation of the first connection being idle.
This problem is solved here by always executing SELECT 1 before we switch to
another connection where we are going to execute SHOW PROCESSLIST. That will
make sure that cleaning of tmp tables after SHOW VARIABLES is done and that
connection will be in the idle state.

Change-Id: I61db46be030aebc5600b91487834c3d88bebab79
Reviewed-by: Jeremy Cole <jeremycole@google.com>