Implement Sniper framework and idle sniper.
This patch implements a sniper which will kill processes deemed to be
behaving badly. The Sniper works by periodically running through the list
of all threads and checking with the loaded Sniper_modules to determine
whether a connection should be killed. The sniper keeps Sniper_modules in
two categories. The global_checks are checks which must all be passed in
order for a connection to be considered for sniping. After passing the
global_checks the connection will be sniped if any of the periodic_checks
(such as the idle snipe module) returns true.

This patch adds the following configuration flags:
  --sniper                        : run the sniper. This is assumed to be true
                                    if any flags that enable modules are
                                    present.
  --sniper-check-period=#         : The number of seconds to wait between
                                    running the sniper
  --sniper-idle-timeout=#         : The number of seconds after which to kill
                                    idle connections
  --sniper-ignore-unauthenticated : Makes the sniper not ignore connections
                                    where the user is not yet authenticated

Furthermore the sniper will always load modules that prevent it from sniping
any connections owned by a user with the SUPER ACL and will never kill any
system thread.

There are 3 currently implemented Sniper_modules:
  Sniper_module_idle            : A periodic check used to check if a threads
                                  time sleeping has expired.
  Sniper_module_priv_ignore     : A global check used to ignore connections
                                  with specified ACLs.
  Sniper_module_unauthenticated : A global check used to ignore connections
                                  where the user has not yet authenticated.

Currently the configuration of the sniper and its associated modules is not
mutable at runtime however the structure of the Sniper should make this a
possibility.

The main Sniper code is in sql/sniper.cc and the code for the Sniper_modules
are in sql/sniper_modules.cc

Change-Id: I701e8d23ae8404793947e700ea7dc04ddd0dc898