Added Connectionless and long_query sniper modules.
Created two new snipers. The connectionless sniper will kill queries where
the client has disconnected and the long_query sniper will kill queries that
take more than a configurable number of seconds.

This change also adds two new flags:
    --sniper-long-query-timeout=# : The number of seconds to allow a query to
                                    run before killing it.
    --sniper-connectionless       : Have the connectionless sniper be run.

The code for these snipers is in sql/sniper_modules.cc.

This change also adds a new mysqltest command: sudden_close. This command
causes the connection to be severed at the file-descriptor level on the
clients side. This is used in the testing of the connectionless sniper.

Change-Id: Ic291fb63171a45a5828d6b4fe08e6a1af7e8e500