Add new audit_log_remote_connections option
Add a new parameter audit_log_remote_connections which will audit log
TCP connections that do not originate from localhost. This is useful to
allow tracking of users connecting from remote systems without polluting
the log with connections from local administrative tasks and similar.

In order to support this logging, Vio had to be extended to populate the
currently unused vio->local structure via getsockname (it is unclear why
this has existed unpopulated in the vio structure).

A connection is considered remote if it:
  * Is via TCP/IP.
  * Has a different source and destination host, and thus is not TCP/IP
    via one of the locally bound IP addresses, such as loopback or any
    address on an Ethernet interface.

Additionally a test was added for the existing audit_log_connections option.

Change-Id: I36e7f551f9af57cb41aa5a2b7d0fb465d6293f61
Reviewed-by: Pavel Ivanov <pivanof@google.com>