Added ability to dynamically change sniper configuration.
This patch adds the ability to modify the sniper configuration
through the use of global sys_vars. You are able to turn the
sniper on and off, set how long it waits between checks, change the
timeouts for long_query and idle snipers and turn them off or on,
and load or unload any of the included modules.

This also modifies the tests so they will explicitly test this
new functionality.

Change-Id: I69580b53d4ab3db9b00f3d74672ee56d0bcf877f