Fix tests to be resilient of test infrastructure
For the "remote connections" test, use a user that is allowed to connect
from any host, rather than the built-in root user. During test setup,
the root account is set up to be able to connect from localhost and the
current host's name based on "select @@hostname", but this may not be
the same name that will be resolved when a connection is made through
that name. Specifically, @@hostname may return the short name while DNS
resolution will always return the FQDN, and the resolved name is what is
used for permissions checks.

Change-Id: I6ce447527580825b294a812f085783c656b2430f
Reviewed-by: Pavel Ivanov <pivanof@google.com>