Add a new infeasible sniper to mysql.
Add a new infeasible sniper that can kill queries that are unlikely to
finish in a reasonable amount of time. It determines the feasibility of a
query by looking at the worst case complexity of the query, as measured
by the product of all the rows multiplied together in the plan for the
query. It will also take into account whether the plan makes use of
temporary tables and the filesort algorithm.

This change introduces several flags:
  --sniper-infeasible-cross-product-rows=#
      The minimum cross product needed for a query to be considered
      possibly infeasible.  This must be set for the sniper to be loaded.

  --sniper-infeasible-max-time=#
      The time, in seconds, we should allow a query that has been judged
      potentially infeasible to run before killing it.  This defaults to
      0.  This is useful because some queries are easier than they appear.

  --sniper-infeasible-secondary-requirements=(NONE |
                                              FILESORT |
                                              TEMPORARY |
                                              FILESORT_OR_TEMPORARY |
                                              FILESORT_AND_TEMPORARY)
      This flag allows you to put additional restrictions on what a plan
      needs to include if it is to be judged infeasible.
      The options are:
      NONE                   : There are no additional requirements on
                               the plan.
      FILESORT               : The plan must include a filesort for it
                               to be judged infeasible.
      TEMPORARY              : The plan must use temporary tables for it
                               to be judged infeasible.
      FILESORT_AND_TEMPORARY : The plan must include a filesort and use
                               a temporary table for it to be judged
                               infeasible.
      FILESORT_OR_TEMPORARY  : The plan must either use temporary tables,
                               include a filesort, or both for it to be
                               judged infeasible.

This patch includes changes to SELECT_LEX_UNIT in order to obtain the
data on the number of lines scanned.  This data about the complexity and
uses of filesort and temporary tables is cached by the lex object and
only calculated once. All three new settings are dynamically configurable.

Change-Id: I868a68a9d21fa642633404a4991b54cd98a10b6a