Adds the ability to configure the sniper on a per user basis.
This change includes a new syntax feature to GRANT command. It adds these
new 'with_option's:
  SNIPER_IDLE_TIMEOUT #
    Sets the idle timeout for the user to the given value. A value of
    0 indicates that the idle sniper is disabled for this user.
  SNIPER_LONG_QUERY_TIMEOUT #
    Sets the long query timeout for the user to the given value. A value
    of 0 indicates the long query sniper is disabled for this user.
  SNIPER_KILL_CONNECTIONLESS ENABLE
    Makes the connectionless sniper kill any connectionless queries made
    by this user.
  SNIPER_KILL_CONNECTIONLESS DISABLE
    Makes the connectionless sniper ignore this user.
  SNIPER_INFEASIBLE_CROSS_PRODUCT_ROWS #
    Sets the minimum cross product needed for a query to be considered
    infeasible for this user to the given value. A value of 0 indicates
    the infeasible sniper is disabled for this user.
  SNIPER_INFEASIBLE_MAX_TIME #
    Sets the time we will allow a potentially infeasible query to finish
    in before killing it for this user.
  SNIPER_INFEASIBLE_SECONDARY_REQUIREMENTS (NONE | FILESORT | TEMPORARY |
                                            FILESORT_OR_TEMPORARY |
                                            FILESORT_AND_TEMPORARY)
    Sets the secondary requirements for the infeasible sniper to the
    associated value for this user.

All these options can also be set to DEFAULT, which causes the associated
sniper modules to use the system default values for sniping decisions. As
with user resource limits in cases where the same setting is specified
more than once in the same command the value at the end is chosen over
earlier ones.

These can be used, for example as:

  GRANT USAGE ON *.* TO 'user1'@'host1'
      WITH SNIPER_IDLE_TIMEOUT 10
           SNIPER_LONG_QUERY_TIMEOUT 0
           SNIPER_INFEASIBLE_CROSS_PRODUCT_ROWS DEFAULT
           SNIPER_KILL_CONNECTIONLESS ENABLE;

These settings are stored persistently in the new mysql.sniper_settings
table, which has a column for each of these settings. This table can be
manually modified in the same manner as the mysql.user table and FLUSH
PRIVILEGES must be run to make chages visible within the server. Renaming
and dropping users with custom sniper settings is handled automatically
by the server.

All of these settings, and also the settings for user resources can be set
using anything that returns a single value (except for sub-queries). This
means that the following is a valid command:

  GRANT USAGE ON *.* to 'user1'@'host1'
    WITH SNIPER_IDLE_TIMEOUT CAST(9990 * RAND() + 10 as INT);

Finally this change makes it so that modules are never fully removed
from the sniper. This means that all user specific settings will still
be followed even if the default setting is to ignore the module.

Change-Id: Ibfa58d3f7d5077d4a98ffcb2a40ecb4891da7379