Add support for replicating from Google MySQL 5.1
In order to support replicating from Google MySQL 5.1, the following
changes were required (in addition to various pre-requisite fixes to the
parsing of event headers with the size larger than the standard one):

  * The Group ID and Checksum are extracted from the event header, if
    present (based on a check of the event's header size) and stored in
    thread variables.
  * Any BEGIN events are converted to GTID events using the extracted
    Group ID.
  * Any Query events outside of the transaction are preceded with GTID
    event.
  * All other queries and events are unmodified.

This allows MariaDB 10.0 to see Google MySQL 5.1 Group IDs as GTIDs.
MariaDB internally makes some other modifications to the received
stream, which are unaffected by this change.

Nothing is yet done with the extracted Checksum information.

Note: When slave first connects to master it always sends Rotate event and
Format_description event. After that data in Format_description event already
defines whether checksums are used by master or not. And they are considered
not to be used by MySQL 5.1 master (as MariaDB has different storage format for
checksums). But the first Rotate event is apparently always checksummed by
MariaDB and never checksummed by MySQL 5.1, and that introduces incompatibility.
Disabling checking of the checksum on the first Rotate event allows MariaDB
to replicate from MySQL 5.1.

Another note: Sometimes MariaDB uses its own Format_description_log_event to
decode Rotate_log_event instead of master's one when replicating from MySQL. I
failed to udnerstand the reason why and the set of events when that happens, but
the consequence is that Rotate_log_event strips last 4 characters from the
master log filename. That filename is then sent to master in the semi-sync acks.
Master don't understand them and doesn't accept such acks.

This commit adds a hack that will allow to mitigate the problem. It's a dirty
hack, but it's needed only till we complete migration to MariaDB. So I think
we'll be fine living with it until it could be deleted.

Also this makes MariaDB understand semi-sync acks coming from MySQL 5.1 when
MariaDB is a master.

Change-Id: Iadd9fca2b8c6f0b474bb9eee7e674f2014f4b78d